requirements.txt
streamlit>=1.22.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=9.5.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
langchain-core>=0.1.10
chromadb>=0.4.18
openai>=1.1.1
regex>=2023.6.3

# Optional dependencies for better PDF OCR
PyMuPDF>=1.22.5
pdf2image>=1.16.3
# poppler-utils (system dependency for pdf2image)